+++ lib/pure/uri.nim
@@ -50,12 +50,6 @@ import strutils, parseutils, base64
 include includes/decode_helpers
 
 
-
-when (NimMajor, NimMinor, NimPatch) >= (1, 3, 5):
-  type
-    IpKind* = enum
-      IPv4, IPv6
-
 type
   Url* = distinct string
 
@@ -63,7 +57,7 @@ type
     scheme*, username*, password*: string
     hostname*, port*, path*, query*, anchor*: string
     opaque*: bool
-    kind: IpKind # not expose it for compatibility.
+    isIpv6: bool # not expose it for compatibility.
 
 proc encodeUrl*(s: string, usePlus = true): string =
   ## Encodes a URL according to RFC3986.
@@ -178,7 +172,7 @@ proc parseAuthority(authority: string, result: var Uri) =
         inPort = true
     of '[':
       inIPv6 = true
-      result.kind = IPv6
+      result.isIpv6 = true
     of ']':
       inIPv6 = false
     else:
@@ -217,28 +211,27 @@ proc initUri*(): Uri =
   result = Uri(scheme: "", username: "", password: "", hostname: "", port: "",
                 path: "", query: "", anchor: "")
 
-proc initUri*(kind: IpKind): Uri {.since: (1, 3, 5).} =
+proc initUri*(isIpv6: bool): Uri {.since: (1, 3, 5).} =
   ## Initializes a URI with ``scheme``, ``username``, ``password``,
-  ## ``hostname``, ``port``, ``path``, ``query``, ``anchor`` and ``kind``.
+  ## ``hostname``, ``port``, ``path``, ``query``, ``anchor`` and ``isIpv6``.
   ##
   ## **See also:**
   ## * `Uri type <#Uri>`_ for available fields in the URI type
   runnableExamples:
-    var uriIpv6: Uri
-
-    let org = "udp://[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8080"
-    parseUri(org, uriIpv6)
-    doAssert uriIpv6.hostname == "2001:0db8:85a3:0000:0000:8a2e:0370:7334"
-    doAssert $uriIpv6 == org
+    var uri2 = initUri(isIpv6 = true)
+    uri2.scheme = "tcp"
+    uri2.hostname = "2001:0db8:85a3:0000:0000:8a2e:0370:7334"
+    uri2.port = "8080"
+    assert $uri2 == "tcp://[2001:0db8:85a3:0000:0000:8a2e:0370:7334]:8080"
   result = Uri(scheme: "", username: "", password: "", hostname: "", port: "",
-                path: "", query: "", anchor: "", kind: kind)
+                path: "", query: "", anchor: "", isIpv6: isIpv6)
 
 proc resetUri(uri: var Uri) =
   for f in uri.fields:
     when f is string:
       f.setLen(0)
-    elif f is IpKind:
-      f = IPv4
+    elif f is bool:
+      f = false
     else:
       f = false
 
@@ -479,17 +472,15 @@ proc `$`*(u: Uri): string =
       result.add(u.password)
     result.add("@")
   if u.hostname.endsWith('/'):
-    case u.kind
-    of IpKind.IPv4:
-      result.add(u.hostname[0 .. ^2])
-    of IpKind.IPv6:
+    if u.isIpv6:
       result.add("[" & u.hostname[0 .. ^2] & "]")
+    else:
+      result.add(u.hostname[0 .. ^2])
   else:
-    case u.kind
-    of IpKind.IPv4:
-      result.add(u.hostname)
-    of IpKind.Ipv6:
+    if u.isIpv6:
       result.add("[" & u.hostname & "]")
+    else:
+      result.add(u.hostname)
   if u.port.len > 0:
     result.add(":")
     result.add(u.port)